+++ vaadin-date-range-picker-demo/src/main/java/software/xdev/vaadin/daterange_picker/example/customized/CustomDateRanges.java
@@ -1,6 +1,6 @@
 package software.xdev.vaadin.daterange_picker.example.customized;
 
-import software.xdev.vaadin.daterange_picker.buisness.SimpleDateRanges;
+import software.xdev.vaadin.daterange_picker.business.SimpleDateRanges;
 
 public class CustomDateRanges
 {
