+++ vaadin-date-range-picker/src/main/java/software/xdev/vaadin/daterange_picker/ui/DateRangePickerOverlay.java
@@ -42,9 +42,9 @@ import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.data.binder.HasItems;
 import com.vaadin.flow.shared.Registration;
 
-import software.xdev.vaadin.daterange_picker.buisness.DateRange;
-import software.xdev.vaadin.daterange_picker.buisness.DateRangeModell;
-import software.xdev.vaadin.daterange_picker.buisness.DateRangeResult;
+import software.xdev.vaadin.daterange_picker.business.DateRange;
+import software.xdev.vaadin.daterange_picker.business.DateRangeModell;
+import software.xdev.vaadin.daterange_picker.business.DateRangeResult;
 
 
 /**
